--- src/components/base-input/base-input.types.ts
@@ -1,7 +1,0 @@
-import React from "react";
-
-export interface BaseInputProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {
-  error?: string;
-  label: string;
-}
